backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pydantic
python-multipart
httpx
